Refactor: Change permissionUtil to abstract class
